refactor(uuid): move UUID handling to separate module and clean up imports
